refactor(ui): extract paced markdown helpers to dedicated module
Slice 11b.1 step 3/14 of the message-part.tsx split. Moves the
24-ms paced text reveal hook (`createPacedValue`), the path helpers
(`isAbsoluteLikePath`, `joinWorkspacePath`, `relativizeProjectPath`,
`getDirectory`), and the `<MessageMarkdown>` / `<PacedMarkdown>`
wrappers into a new `message-part-markdown.tsx` module.

These helpers are consumed by the text / reasoning renderers (paced
streaming), the assistant message display (markdown body), and a
handful of tool renderers that show paths. Lifting them out lets each
consumer import the markdown stack directly without dragging the
full message-part dispatcher's transitive graph.

Behavior is unchanged: same 24-ms cadence, same 8-char whitespace
lookahead, same cleanup semantics on signal change / unmount. The
`useData()` access inside `<MessageMarkdown>` and `getDirectory()` is
preserved at the same call sites — these helpers must be invoked
inside Solid render scopes, never at module load.

message-part.tsx 2399 → 2263 lines (-136). Typecheck GREEN, 478 pass /
3 baseline fail unchanged.
